moved bookings and coupons in navigation
