+++ src/components/Navbar.vue
@@ -153,7 +153,6 @@
 import { mapActions, mapGetters } from "vuex";
 import ToastService from "@/services/ToastService";
 import ApiAuthService from "@/services/api/ApiAuthService";
-import { RolePermission } from "@/entities/role";
 import ApiTenantService from "@/services/api/ApiTenantService";
 
 export default {
@@ -199,6 +198,11 @@ export default {
             icon: "mdi-account-group-outline",
             interfaceName: "roles",
           },
+        ],
+      },
+      {
+        header: "Buchungsplattform",
+        pages: [
           {
             title: "Buchungen",
             link: "buchungen",
@@ -211,11 +215,6 @@ export default {
             icon: "mdi-ticket-percent-outline",
             interfaceName: "coupons",
           },
-        ],
-      },
-      {
-        header: "Buchungsplattform",
-        pages: [
           {
             title: "Veranstaltungsorte",
             link: "veranstaltungsorte",
