+++ src/views/Bookings.vue
@@ -102,9 +102,9 @@ import Fuse from "fuse.js";
 import AdminLayout from "@/layouts/Admin.vue";
 import { mapActions, mapGetters } from "vuex";
 import ApiBookingService from "@/services/api/ApiBookingService";
-import BookingEdit from "@/components/Booking/BookingEdit";
-import BookingDeleteConformationDialog from "@/components/Booking/BookingDeleteConformationDialog";
-import BookingRejectConformationDialog from "@/components/Booking/BookingRejectConformationDialog";
+import BookingEdit from "@/components/Booking/BookingEdit.vue";
+import BookingDeleteConformationDialog from "@/components/Booking/BookingDeleteConformationDialog.vue";
+import BookingRejectConformationDialog from "@/components/Booking/BookingRejectConformationDialog.vue";
 import ApiBookablesService from "@/services/api/ApiBookablesService";
 import BookingPermissionService from "@/services/permissions/BookingPermissionService";
 import BookingDetails from "@/components/Booking/BookingDetails.vue";
